Added Censys functionality
Moved Censys search function into its own file
